Re-org structure and import shared utilities
