+++ src/idi_company_info/company_by_cusip_pipeline.py
