+++ src/idi_company_info/types.py
